+++ src/cli.rs
@@ -66,7 +66,8 @@ pub fn main() -> Result<()> {
         .subcommand(cmd::config::new())
         .subcommand(cmd::script::new())
         .subcommand(cmd::checkout::new())
-        .subcommand(cmd::vendor::new());
+        .subcommand(cmd::vendor::new())
+        .subcommand(cmd::fusesoc::new());
 
     // Add the `--debug` option in debug builds.
     let app = if cfg!(debug_assertions) {
@@ -142,12 +143,18 @@ pub fn main() -> Result<()> {
 
     // Resolve the dependencies if the lockfile does not exist or is outdated.
     let locked = match matches.subcommand() {
-        Some((command, _)) =>
-        {
+        Some((command, matches)) => {
             #[allow(clippy::unnecessary_unwrap)]
-            if command == "update" || locked_existing.is_none() {
+            // execute pre-dependency-fetch commands
+            if command == "fusesoc" && matches.get_flag("single") {
+                return cmd::fusesoc::run_single(&sess, matches);
+            } else if command == "update" || locked_existing.is_none() {
                 if manifest.frozen {
-                    return Err(Error::new(format!("Refusing to update dependencies because the package is frozen. Remove the `frozen: true` from {:?} to proceed; there be dragons.", manifest_path)));
+                    return Err(Error::new(format!(
+                        "Refusing to update dependencies because the package is frozen.
+                        Remove the `frozen: true` from {:?} to proceed; there be dragons.",
+                        manifest_path
+                    )));
                 }
                 debugln!("main: lockfile {:?} outdated", lock_path);
                 let res = DependencyResolver::new(&sess);
@@ -251,6 +258,7 @@ pub fn main() -> Result<()> {
         Some(("checkout", matches)) => cmd::checkout::run(&sess, matches),
         Some(("update", _)) => Ok(()),
         Some(("vendor", matches)) => cmd::vendor::run(&sess, matches),
+        Some(("fusesoc", matches)) => cmd::fusesoc::run(&sess, matches),
         Some((plugin, matches)) => execute_plugin(&sess, plugin, matches.get_many::<OsString>("")),
         _ => Ok(()),
     }